Add tests

+++ contracts/assets/EtherToken.sol
@@ -14,7 +14,7 @@ contract EtherToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Ether Token";
     string public constant symbol = "ETH-T";
-    uint8 public constant decimals = 18;
+    uint public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // EVENTS
