+++ contracts/assets/PreminedAsset.sol
@@ -11,7 +11,7 @@ contract PreminedAsset is Asset {
 
     // METHODS
 
-    function PreminedAsset(string _name, string _symbol, uint8 _decimals, uint256 _amount)
+    function PreminedAsset(string _name, string _symbol, uint _decimals, uint256 _amount)
         Asset(_name, _symbol, _decimals)
     {
         balances[msg.sender] = balances[msg.sender].add(_amount);
